requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
boto3==1.24.89
botocore==1.27.89
cachetools==5.2.0
cffi==1.15.1
constantly==15.1.0
cryptography==38.0.1
Deprecated==1.2.13
Django==4.1.2
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
google-auth==2.12.0
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
jmespath==1.0.1
msgpack==1.0.4
packaging==21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
rsa==4.9
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
zope.interface==5.5.0
